[#6.0.x] - validator tests
