+++ tests/integration/Filter/Validation/Validator/Alnum/ValidateCest.php
@@ -15,10 +15,38 @@ namespace Phalcon\Tests\Integration\Filter\Validation\Validator\Alnum;
 
 use IntegrationTester;
 use Phalcon\Filter\Validation;
+use Phalcon\Filter\Validation\Exception;
 use Phalcon\Filter\Validation\Validator\Alnum;
+use stdClass;
 
 class ValidateCest
 {
+    /**
+     * Tests Phalcon\Filter\Validation\Validator\Alnum :: validate() - empty
+     *
+     * @param IntegrationTester $I
+     *
+     * @return void
+     * @throws Exception
+     *
+     * @author Phalcon Team <team@phalcon.io>
+     * @since  2023-08-03
+     */
+    public function filterValidationValidatorAlnumValidateEmpty(IntegrationTester $I)
+    {
+        $I->wantToTest("Validation\Validator\Alnum - validate() - empty");
+
+        $validation = new Validation();
+        $validator  = new Alnum(['allowEmpty' => true,]);
+        $validation->add('name', $validator);
+        $entity = new stdClass();
+        $entity->name = '';
+
+        $validation->bind($entity, []);
+        $result = $validator->validate($validation, 'name');
+        $I->assertTrue($result);
+    }
+
     /**
      * Tests Phalcon\Filter\Validation\Validator\Alnum :: validate() - single field
      *
@@ -30,11 +58,7 @@ class ValidateCest
         $I->wantToTest("Validation\Validator\Alnum - validate() - single field");
 
         $validation = new Validation();
-
-        $validation->add(
-            'name',
-            new Alnum()
-        );
+        $validation->add('name', new Alnum());
 
         $messages = $validation->validate(
             [
@@ -42,10 +66,9 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
-            0,
-            $messages->count()
-        );
+        $expected = 0;
+        $actual   = $messages->count();
+        $I->assertSame($expected, $actual);
 
         $messages = $validation->validate(
             [
@@ -53,10 +76,9 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
-            1,
-            $messages->count()
-        );
+        $expected = 1;
+        $actual   = $messages->count();
+        $I->assertSame($expected, $actual);
     }
 
     /**
@@ -67,7 +89,7 @@ class ValidateCest
      */
     public function filterValidationValidatorAlnumValidateMultipleField(IntegrationTester $I)
     {
-        $I->wantToTest("Validation\Validator\Alnum - validate() - multiple field");
+        $I->wantToTest("Validation\Validator\Alnum - validate() - multiple fields");
 
         $validation = new Validation();
 
@@ -90,10 +112,6 @@ class ValidateCest
             $al
         );
 
-        codecept_debug($validation);
-        codecept_debug($al);
-
-
         $messages = $validation->validate(
             [
                 'name' => 'SomeValue123',
@@ -101,11 +119,9 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
-            0,
-            $messages->count()
-        );
-
+        $expected = 0;
+        $actual   = $messages->count();
+        $I->assertSame($expected, $actual);
 
         $messages = $validation->validate(
             [
@@ -114,16 +130,13 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
-            1,
-            $messages->count()
-        );
+        $expected = 1;
+        $actual   = $messages->count();
+        $I->assertSame($expected, $actual);
 
-        $I->assertEquals(
-            $validationMessages['name'],
-            $messages->offsetGet(0)->getMessage()
-        );
-
+        $expected = $validationMessages['name'];
+        $actual   = $messages->offsetGet(0)->getMessage();
+        $I->assertSame($expected, $actual);
 
         $messages = $validation->validate(
             [
@@ -132,19 +145,16 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
-            2,
-            $messages->count()
-        );
+        $expected = 2;
+        $actual   = $messages->count();
+        $I->assertSame($expected, $actual);
 
-        $I->assertEquals(
-            $validationMessages['name'],
-            $messages->offsetGet(0)->getMessage()
-        );
+        $expected = $validationMessages['name'];
+        $actual   = $messages->offsetGet(0)->getMessage();
+        $I->assertSame($expected, $actual);
 
-        $I->assertEquals(
-            $validationMessages['type'],
-            $messages->offsetGet(1)->getMessage()
-        );
+        $expected = $validationMessages['type'];
+        $actual   = $messages->offsetGet(1)->getMessage();
+        $I->assertSame($expected, $actual);
     }
 }
