+++ tests/integration/Filter/Validation/Validator/Between/ValidateCest.php
@@ -15,15 +15,49 @@ namespace Phalcon\Tests\Integration\Filter\Validation\Validator\Between;
 
 use IntegrationTester;
 use Phalcon\Filter\Validation;
+use Phalcon\Filter\Validation\Exception;
 use Phalcon\Filter\Validation\Validator\Between;
 use Phalcon\Messages\Message;
 use Phalcon\Messages\Messages;
+use stdClass;
 
 /**
  * Class ValidateCest
  */
 class ValidateCest
 {
+    /**
+     * Tests Phalcon\Filter\Validation\Validator\Between :: validate() - empty
+     *
+     * @param IntegrationTester $I
+     *
+     * @return void
+     * @throws Exception
+     *
+     * @author Phalcon Team <team@phalcon.io>
+     * @since  2023-08-03
+     */
+    public function filterValidationValidatorBetweenValidateEmpty(IntegrationTester $I)
+    {
+        $I->wantToTest("Validation\Validator\Between - validate() - empty");
+
+        $validation = new Validation();
+        $validator  = new Between(
+            [
+                'allowEmpty' => true,
+                'minimum'    => 1,
+                'maximum'    => 3,
+            ]
+        );
+        $validation->add('price', $validator);
+        $entity = new stdClass();
+        $entity->price = '';
+
+        $validation->bind($entity, []);
+        $result = $validator->validate($validation, 'price');
+        $I->assertTrue($result);
+    }
+
     /**
      * Tests Phalcon\Filter\Validation\Validator\Between :: validate() - single field
      *
@@ -66,13 +100,11 @@ class ValidateCest
 
         $I->assertEquals($expected, $messages);
 
-
         $messages = $validation->validate(
             []
         );
 
         $I->assertEquals($expected, $messages);
-
 
         $messages = $validation->validate(
             [
@@ -80,7 +112,7 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
+        $I->assertSame(
             0,
             $messages->count()
         );
@@ -130,7 +162,7 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
+        $I->assertSame(
             0,
             $messages->count()
         );
@@ -143,12 +175,12 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
+        $I->assertSame(
             1,
             $messages->count()
         );
 
-        $I->assertEquals(
+        $I->assertSame(
             $validationMessages['amount'],
             $messages->offsetGet(0)->getMessage()
         );
@@ -161,17 +193,17 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
+        $I->assertSame(
             2,
             $messages->count()
         );
 
-        $I->assertEquals(
+        $I->assertSame(
             $validationMessages['amount'],
             $messages->offsetGet(0)->getMessage()
         );
 
-        $I->assertEquals(
+        $I->assertSame(
             $validationMessages['price'],
             $messages->offsetGet(1)->getMessage()
         );
@@ -220,12 +252,10 @@ class ValidateCest
 
         $I->assertEquals($expected, $messages);
 
-
         $I->assertEquals(
             $validation->validate([]),
             $messages
         );
-
 
         $messages = $validation->validate(
             [
@@ -233,7 +263,7 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
+        $I->assertSame(
             0,
             $messages->count()
         );
