+++ tests/integration/Filter/Validation/Validator/StringLength/Min/GetTemplateCest.php
@@ -31,16 +31,16 @@ class GetTemplateCest
         $validator = new Min();
 
         $expected = 'Field :field must be at least :min characters long';
-        $I->assertEquals($expected, $validator->getTemplate(), 'Default template message for Min');
+        $I->assertSame($expected, $validator->getTemplate(), 'Default template message for Min');
 
         $validator->setTemplate('');
 
         $expected = 'The field :field is not valid for ' . Min::class;
-        $I->assertEquals($expected, $validator->getTemplate(), 'Default template message');
+        $I->assertSame($expected, $validator->getTemplate(), 'Default template message');
 
         $expected = 'New custom template';
         $validator->setTemplate($expected);
 
-        $I->assertEquals($expected, $validator->getTemplate(), 'New template message');
+        $I->assertSame($expected, $validator->getTemplate(), 'New template message');
     }
 }
