+++ tests/integration/Filter/Validation/Validator/StringLength/Min/ValidateCest.php
@@ -15,10 +15,38 @@ namespace Phalcon\Tests\Integration\Filter\Validation\Validator\StringLength\Min
 
 use IntegrationTester;
 use Phalcon\Filter\Validation;
+use Phalcon\Filter\Validation\Exception;
 use Phalcon\Filter\Validation\Validator\StringLength\Min;
+use stdClass;
 
 class ValidateCest
 {
+    /**
+     * Tests Phalcon\Filter\Validation\Validator\Min :: validate() - empty
+     *
+     * @param IntegrationTester $I
+     *
+     * @return void
+     * @throws Exception
+     *
+     * @author Phalcon Team <team@phalcon.io>
+     * @since  2023-08-03
+     */
+    public function filterValidationValidatorMinValidateEmpty(IntegrationTester $I)
+    {
+        $I->wantToTest("Validation\Validator\Min - validate() - empty");
+
+        $validation = new Validation();
+        $validator  = new Min(['allowEmpty' => true,]);
+        $validation->add('name', $validator);
+        $entity = new stdClass();
+        $entity->name = '';
+
+        $validation->bind($entity, []);
+        $result = $validator->validate($validation, 'name');
+        $I->assertTrue($result);
+    }
+
     /**
      * Tests Phalcon\Filter\Validation\Validator\StringLength :: validate() - single
      * field
@@ -48,7 +76,7 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
+        $I->assertSame(
             0,
             $messages->count()
         );
@@ -60,7 +88,7 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
+        $I->assertSame(
             0,
             $messages->count()
         );
@@ -96,7 +124,7 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
+        $I->assertSame(
             1,
             $messages->count()
         );
@@ -108,7 +136,7 @@ class ValidateCest
             ]
         );
 
-        $I->assertEquals(
+        $I->assertSame(
             0,
             $messages->count()
         );
