--- tests/integration/Filter/Validation/Validator/ValidateCest.php
@@ -1,34 +1,0 @@
-<?php
-
-/**
- * This file is part of the Phalcon Framework.
- *
- * (c) Phalcon Team <team@phalcon.io>
- *
- * For the full copyright and license information, please view the LICENSE.txt
- * file that was distributed with this source code.
- */
-
-declare(strict_types=1);
-
-namespace Phalcon\Tests\Integration\Filter\Validation\Validator;
-
-use IntegrationTester;
-
-/**
- * Class ValidateCest
- */
-class ValidateCest
-{
-    /**
-     * Tests Phalcon\Filter\Validation\Validator :: validate()
-     *
-     * @author Phalcon Team <team@phalcon.io>
-     * @since  2018-11-13
-     */
-    public function filterValidationValidatorValidate(IntegrationTester $I)
-    {
-        $I->wantToTest('Validation\Validator - validate()');
-        $I->skipTest('Need implementation');
-    }
-}
